refactor: improve route matching logic and introduce new route pattern handling
